Mannual ban added without effect

+++ src/routes/api/team/[team_id]/delete_team/+server.ts
@@ -1,6 +1,6 @@
 import {json} from '@sveltejs/kit'
 import Teams from '$lib/server/database/teams'
-
+import Contest from '$lib/server/database/contest.js';
 
 export async function POST({ locals,params,request}) {
    
@@ -12,16 +12,28 @@ export async function POST({ locals,params,request}) {
     const user_id = locals.user.id;
 
     // check leader access
-    const leader_id_response = await Teams.get_leader_id(team_id);
+    const leader_id_response = await Teams.get_team_info(team_id);
 
     if(leader_id_response.error) return json({success: false, message: "Something went wrong"});
     if(leader_id_response.data.length == 0) return json({success: false, message: "Team not found"});
     if(leader_id_response.data[0].leader_id != user_id) return json({success: false, message: "You are not the leader of this team"});
     
 
+    // check the contest status
+    const contest_id = leader_id_response.data[0].contest_id;
+    const contest_status = await Contest.get_contest_status(contest_id);
+
+    if(contest_status.error) return json({success: false, message: "Something went wrong"});
+    if(contest_status.data.length == 0) return json({success: false, message: "Contest not found"});
+
+    const status = contest_status.data[0].status;
+
+    if(!status.status || status.status == 'finished') return json({success: false, message: "Can not delete team after the contest has finished"});
+
+
     // delete team
     const delete_response = await Teams.delete_team(team_id);
     if(delete_response.error) return json({success: false, message: "Something went wrong"});
 
-    return json({success: true, message: "Password set successfully"});
+    return json({success: true, message: "Team deleted successfully"});
 }
